Move AuthRolesProvider into the pass-authz-roles module
